Fix reflection issues in UserDefinedInstanceProperty
